(browser-tests) Separate setup and test dirs, change assertion in callout test
Changed assertion in callout test to use what's visible to the user, instead of what happens in the background
