+++ apps/browser-tests/README.md
@@ -1,11 +1,60 @@
-# browser-tests
+# @beabee/browser-tests
 
-Run a test:
+Playwright browser tests for the Beabee frontend. Tests run against Firefox (Desktop) by default. A setup project handles authentication before the test suite runs.
+
+## Prerequisites
+
+A running Beabee instance (default: `http://localhost:4002/`). Override by changing `playwright.config.ts`
+
+## Running tests
+
 ```sh
+# Run all tests (headless, with trace on failure)
+yarn workspace @beabee/browser-tests test
+
+# Run with interactive UI
 yarn workspace @beabee/browser-tests test:ui
 ```
 
-Record tests:
+## Recording new tests
+
 ```sh
 yarn workspace @beabee/browser-tests playwright codegen localhost:3000
 ```
+
+## Setup
+
+Before tests run, a setup project (`src/setup/auth.setup.ts`) authenticates two users and saves their browser storage state to `.auth/` (git-ignored):
+
+| File | User | Credentials source |
+| ---- | ---- | ------------------ |
+| `.auth/member.json` | Non-admin member | `rateLimitedTestUser` from `@beabee/test-utils/test-data` |
+| `.auth/admin.json` | Admin | `testUser` from `@beabee/test-utils/test-data` |
+
+Each test file opts in to a stored auth state via `test.use({ storageState: ... })`. Tests that need to act as both users open additional browser contexts with `browser.newContext({ storageState: ... })`.
+
+The setup runs automatically before the test suite — no manual step required.
+
+## Project structure
+
+```
+src/
+  setup/
+    auth.setup.ts       # Authenticates admin + non-admin users, stores auth state to .auth/
+    auth-states.ts      # Exports paths to stored auth state files
+  tests/
+    *.spec.ts           # Test files
+.auth/                  # Stored authentication state (git-ignored)
+playwright.config.ts
+```
+
+## Configuration
+
+| Option                | Default                  | Description                                                               |
+| --------------------- | ------------------------ | ------------------------------------------------------------------------- |
+| `PLAYWRIGHT_BASE_URL` | `http://localhost:4002/` | Base URL for the running instance                                         |
+| `workers`             | `4`                      | Parallel file-level concurrency (tests within a file always run serially) |
+| `trace`               | `retain-on-failure`      | Playwright trace mode                                                     |
+| `screenshot`          | `only-on-failure`        | Screenshot capture mode                                                   |
+
+Reports are written to `playwright-report/` and test artifacts to `test-results/`.
